Move NSURLRequest swizzling to own file
